requirement.txt
pandas 
plotly_express 
<<<<<<< HEAD
streamlit
=======
streamlit
>>>>>>> 102185136d615308ac6686628dc09fbf1bd33165
